Reaction Time Detection and Log
Implemented a simple system whereby, when the strobe is activated (using CB_MODE button), a timer is started. When user presses space bar, the timer will stop and log the reaction time and 'reason' - which will be our type of visual cue. In reality this can be adapted to whatever the user is expected to do upon seeing the visual cue.