Added three visual cues
Three visual cues were added, all combined with a countdown on the dashboard before the vehicle will switch to 'AI' mode. These are:
1. Auditory 'beep'
2. Light flash
3. Countdown only

The three cues are in three different 'base' files. These must be dragged into the 'Manager' object to change the cue before starting the experience.
At present, WP4 will cause the cues to function. This needs to be changed to wherever required. Right now, WP4 is too close to the accident and the vehicle 'bugs out' before transferring to MDV mode.